Split SearchViewListener to ModuleViewListener (task #809)
