+++ config/bootstrap.php
@@ -55,6 +55,7 @@ use App\Event\Plugin\Menu\View\ApplicationMenuListener;
 use App\Event\Plugin\Menu\View\DashboardMenuListener;
 use App\Event\Plugin\Menu\View\ModuleIndexListener;
 use App\Event\Plugin\Menu\View\ModuleMenuListener;
+use App\Event\Plugin\Menu\View\ModuleViewListener;
 use App\Event\Plugin\Menu\View\SearchViewListener;
 use App\Event\Plugin\Search\Model\ChildListItemsListener;
 use App\Event\Plugin\Search\Model\ReportsListener;
@@ -283,6 +284,7 @@ EventManager::instance()->on(new ApplicationMenuListener());
 EventManager::instance()->on(new DashboardMenuListener());
 EventManager::instance()->on(new ModuleMenuListener());
 EventManager::instance()->on(new ModuleIndexListener());
+EventManager::instance()->on(new ModuleViewListener());
 EventManager::instance()->on(new SearchViewListener());
 EventManager::instance()->on(new AccessMenuListener());
 EventManager::instance()->on(new RelatedActionListener());
